ja roda os testes

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: heolivei <heolivei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/05 15:02:24 by heolivei          #+#    #+#             */
-/*   Updated: 2023/01/09 22:21:09 by heolivei         ###   ########.fr       */
+/*   Updated: 2023/01/10 22:09:02 by heolivei         ###   ########.fr       */
 /*   Created: 2022/12/08 15:37:47 by heolivei          #+#    #+#             */
 /*   Updated: 2023/01/07 13:17:46 by heolivei         ###   ########.fr       */
 /*                                                                            */
@@ -29,7 +29,7 @@ char	*ft_brute_line(int fd, char *brute_line);
 char	*ft_liquid_line(char *brute_line);
 char	*ft_rest_line(char *brute_line);
 
-int		ft_strlen(char *s);
+size_t	ft_strlen(char *s);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strchr(char *s, int c);
